Fixed checker pieces positions, added cylinders for board position tracking
The checkers are supposed to be arranged so that the lower left corner of each player's side has a checker in it, so I fixed that. Also, I added cylinders for each valid position on the board; these could be used for collision events to verify that the player is putting a checker in a valid spot on the board.